+++ orix/tests/io/test_io.py
@@ -73,7 +73,7 @@ class TestGeneralIO:
         with pytest.raises(IOError, match=f"No filename matches '{fname}'."):
             _ = load(fname)
 
-    @pytest.mark.parametrize("temp_file_path", ["ctf"], indirect=["temp_file_path"])
+    @pytest.mark.parametrize("temp_file_path", ["ktf"], indirect=["temp_file_path"])
     def test_load_unsupported_format(self, temp_file_path):
         np.savetxt(temp_file_path, X=np.random.rand(100, 8))
         with pytest.raises(IOError, match=f"Could not read "):
@@ -146,11 +146,12 @@ class TestGeneralIO:
         assert crystal_map2.phases[0].name == expected_phase_name
 
 
+# TODO: Remove after 0.13.0
 def test_loadctf():
-    """Crude test of the ctf loader"""
     z = np.random.rand(100, 8)
     fname = "temp.ctf"
     np.savetxt(fname, z)
 
-    _ = loadctf(fname)
+    with pytest.warns(np.VisibleDeprecationWarning):
+        _ = loadctf(fname)
     os.remove(fname)