+++ CHANGELOG.rst
@@ -6,6 +6,30 @@ All user facing changes to this project are documented in this file. The format 
 on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`__, and this project tries
 its best to adhere to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
+2024-09-03 - version 0.13.0
+===========================
+
+Added
+-----
+- We can now read 2D crystal maps from Channel Text Files (CTFs) using ``io.load()``.
+
+Changed
+-------
+- Phase names in crystal maps read from .ang files with ``io.load()`` now prefer to use
+  the abbreviated "Formula" instead of "MaterialName" in the file header.
+
+Removed
+-------
+- Removed deprecated ``from_neo_euler()`` method for ``Quaternion`` and its subclasses.
+- Removed deprecated argument ``convention`` in ``from_euler()`` and ``to_euler()``
+  methods for ``Quaternion`` and its subclasses. Use ``direction`` instead. Passing
+  ``convention`` will now raise an error.
+
+Deprecated
+----------
+- ``loadang()`` and ``loadctf()`` are deprecated and will be removed in the next minor
+  release. Please use ``io.load()`` instead.
+
 2024-04-21 - version 0.12.1
 ===========================